Refactored Values folder to prepare for RNWeb
Moved types to types.ts and implementation in api.ts to separate implementation and declaration.
